gpui(windows): Fix unwrap panic when monitor goes missing

+++ crates/gpui_linux/src/linux/x11/client.rs
@@ -1557,7 +1557,7 @@ impl LinuxClient for X11Client {
             X11Display::new(
                 &state.xcb_connection,
                 state.scale_factor,
-                u32::from(id) as usize,
+                u64::from(id) as usize,
             )
             .ok()?,
         ))
